requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
pydantic
SQLAlchemy
psycopg2-binary
python-dotenv
